Match attention mask dtype to activations

+++ llm_lite/model/gpt.py
@@ -45,7 +45,12 @@ class DenseGpt(CachedAutoregressiveModel):
             self.token_embedding(token_ids) + self.position_embedding(positions)[None, :, :]
         )
         causal_mask = torch.triu(
-            torch.full((sequence_length, sequence_length), float("-inf"), device=token_ids.device),
+            torch.full(
+                (sequence_length, sequence_length),
+                float("-inf"),
+                device=token_ids.device,
+                dtype=hidden_states.dtype,
+            ),
             diagonal=1,
         )
         for block in self.blocks:
